playable(bgi): final-review minor cleanups (u8 mode-flags, misnomer/doc sweep)
- host_bgi.c/.h: bgi_write_mode_flag_a/b u16→u8 (DGROUP 0x541f/0x5420 are 1 byte each; values 2 and 1 fit in u8)
- level.c: fix dead load_palette extern signature in #ifdef BUMPY_COPY_PROTECTION block (void→u16 src_off, u16 src_seg)
- tools/screens_oracle.py: replace ~10 "upload_vga_dac_palette DAC writer/dispatch" prose sites with corrected "wait_vretrace_thunk (vsync wait) / vga_dac_upload_from_buffer (genuine DAC writer)"; FN_NAMES[0x9864] and DAC_FNS scoping logic untouched
- src/screens.c:6, screens.h:16, host/host_resource.c:22: stale "upload_vga_dac_palette" range labels → "wait_vretrace_thunk" / "vga_dac_upload_from_buffer"
- docs/reconstruction-fidelity.md: §Task-11 disambiguate host dispatch_by_palette_mode shim from renamed engine wait_vretrace_dispatch; tighten run-on parenthetical at §DAC carve-out

BUMPY.EXE md5 unchanged (cac9ff236a832284fec6fafff2d8602b); playable links with 0 -wx warnings.

+++ src/host/host_bgi.c
@@ -57,8 +57,8 @@ static u8 host_bgi_page_palette[2][48];
  *
  * Declared here because these globals are only needed under #ifdef BUMPY_PLAYABLE:
  * the default BUMPY.EXE NOP stub for fun_7b4a never reaches them. */
-u16 bgi_write_mode_flag_a;   /* DGROUP 0x541f */
-u16 bgi_write_mode_flag_b;   /* DGROUP 0x5420 */
+u8 bgi_write_mode_flag_a;   /* DGROUP 0x541f */
+u8 bgi_write_mode_flag_b;   /* DGROUP 0x5420 */
 
 /* ── host_bgi_stage_image_palette ───────────────────────────────────────────
  * Functional equivalent of the BGI VGA palette-stage handler (1ab9:0620).
